modified assignment8 driver code
